publisher/requirements.txt
google-cloud-storage==1.30.0
squid-py==0.9.3